requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.0
Flask-JWT-Extended==4.7.0
flask-cors==6.0.2
bcrypt==5.0.0
marshmallow==4.2.2
python-dotenv==1.2.2
